Moved navigation mixin as ChartNavigationComposition.
