+++ ts/Extensions/Annotations/NavigationBindings.ts
@@ -18,7 +18,7 @@ import type PointerEvent from '../../Core/PointerEvent';
 
 import Annotation from './Annotations.js';
 import Chart from '../../Core/Chart/Chart.js';
-import chartNavigationMixin from '../../Mixins/Navigation.js';
+import chartNavigationMixin from '../../Core/Chart/ChartNavigationComposition.js';
 import F from '../../Core/FormatUtilities.js';
 const { format } = F;
 import H from '../../Core/Globals.js';
